+++ react_app/carwash/src/components/Navigationbar.js
@@ -1,22 +1,25 @@
-import { Nav, Navbar } from "react-bootstrap";
-import { Link } from "react-router-dom/cjs/react-router-dom";
+import React from 'react';
+import { Navbar, Nav } from 'react-bootstrap';
+import { Link } from 'react-router-dom';
 
-const Navigationbar = () => {
-    return (
-        <Navbar bg="light" data-bs-theme="light" >
-            <Navbar.Text style={{color: "black"}}> BnB </Navbar.Text>
-                <Nav fill variant="tabs">
-                <Nav.Item><Nav.Link class="nav-item active" eventKey="1" as={Link} to="/">Home</Nav.Link></Nav.Item>
-                <Nav.Item><Nav.Link class="nav-item" eventKey="2" as={Link} to="/Details">Details</Nav.Link></Nav.Item>
-                <Nav.Item><Nav.Link class="nav-item" eventKey="3" as={Link} to="/Locations">Locations</Nav.Link></Nav.Item>
-                <Nav.Item><Nav.Link class="nav-item" eventKey="4" as={Link} to="/About">About</Nav.Link></Nav.Item>
-                    
-                    
-                    
-                    
-                </Nav>
-        </Navbar>
-    )
-}
+const NavigationBar = ({ cartCount }) => (
+  <Navbar bg="light" expand="lg">
+    <Navbar.Brand as={Link} to="/">B'n B </Navbar.Brand>
+    <Navbar.Toggle aria-controls="basic-navbar-nav" />
+    <Navbar.Collapse id="basic-navbar-nav">
+      <Nav className="me-auto">
+        <Nav.Link as={Link} to="/">Home</Nav.Link>
+        <Nav.Link as={Link} to="/wash-finder">WashFinder</Nav.Link>
+        <Nav.Link as={Link} to="/about">About Us</Nav.Link>
+        <Nav.Link as={Link} to="/support">Support</Nav.Link>
+      </Nav>
+      <Nav className="ms-auto">  {/* Right align this Nav */}
+        <Nav.Link as={Link} to="/cart">
+          <i className="bi-cart"></i> ({cartCount})
+        </Nav.Link>
+      </Nav>
+    </Navbar.Collapse>
+  </Navbar>
+);
 
-export default Navigationbar;
+export default NavigationBar;
