+++ react_app/carwash/src/index.js
@@ -1,9 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './App';
 import reportWebVitals from './reportWebVitals';
+import App from './components/App';
+import './App.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
+import { createRoot } from 'react-dom/client';
+
 
 ReactDOM.render(
   <React.StrictMode>
@@ -12,7 +15,13 @@ ReactDOM.render(
   document.getElementById('root')
 );
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+const container = document.getElementById('root');
+const root = createRoot(container);
+root.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
+);
+
 reportWebVitals();
+
